pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "jamftf"
authors = [
    {name = "Joseph Little"},
]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "requests",
    "jamfpy @ git+https://github.com/thejoeker12/jamfpy@main",
    "python-dotenv>=1.0.0",
]
dynamic = ["version"]

[project.optional-dependencies]
pdf = ["ReportLab>=1.2", "RXP"]
rest = ["docutils>=0.3", "pack ==1.1, ==1.3"]

[project.scripts]
my-script = "my_package.module:function"